Add all images into static directory in preparation for move to Netlify
This commit corrects all of my blog posts from 2011-2014 inclusive so all
images are referenced as "/" and added into the "static" directory. This
effectively makes the site self-hosting.

Note that there is a single post from January, 2015, where I appear to
have lost the images. This commit does not correct this post.

Signed-off-by: Kyle Mestery <mestery@mestery.com>